Add files via upload
With Pr. Anne-Ruiz Gazen’s help, I worked on a subject about the Python package
scikit-fda and its difference with the R’s package fda more commonly used for Functional
Data Analysis. We will see that they can be some differences in the results depending
on which package is used.